tests/wpt/web-platform-tests/websockets/Create-protocols-repeated-case-insensitive.htm

<!DOCTYPE html>
<html>
<head>
    <title>W3C WebSocket API - Create WebSocket - repeated protocols with different case</title>
    <script type="text/javascript" src="/resources/testharness.js"></script>
    <script type="text/javascript" src="/resources/testharnessreport.js"></script>
    <script type="text/javascript" src="websocket.sub.js"></script>
</head>
<body>
    <div id="log"></div>
    <script type="text/javascript">
        test(function () {
            var wsocket;
            assert_throws("SYNTAX_ERR", function () { wsocket = CreateWebSocketWithRepeatedProtocolsCaseInsensitive() });
        }, "W3C WebSocket API - Create WebSocket - Pass a valid URL and an array of protocol strings with repeated values but different case - SYNTAX_ERR is thrown")
    </script>
</body>
</html>
